requirements.txt
tweepy
pandas_profiling
streamlit
sklearn
numpy
transformers
pytorch
wordcloud
matplotlib